fixed some things

+++ fns_and_dsa/shopping_list_manager.py
@@ -5,16 +5,18 @@ def display_menu():
     print("3. View List")
     print("4. Exit")
 
-def display_items():
+def display_items(shopping_list):
     print("All Items")
     print("-----------")
     for idx, item in enumerate(shopping_list, start=1):
         print(f"{idx}. {item}")
 
 def main():
+    shopping_list = []
+    
     while True:
         display_menu()
-        choice = input("Enter your choice: ")
+        choice = input("Enter your choice (by number): ")
 
         if choice == '1':
             # Prompt for and add an item
@@ -27,29 +29,31 @@ def main():
                 
         elif choice == '2':
             # Prompt for and remove an item
-            if not len(shopping_list) == 0:
+            if shopping_list:
                 display_items()
             else:
                 print("Item list is empty.")
-                break
+                continue
             try:
                 remove = int(input("Remove item (by number): "))
-                if 1 <= remove <= len(shopping_list):
-                    remove_index = remove - 1
-                    shopping_list.pop(remove_index)
-                    print("Item successfully removed.")
-                else:
-                    print("Invalid number. Please enter a valid item number.")
+                if remove:
+                    if 1 <= remove <= len(shopping_list):
+                        remove_index = remove - 1
+                        shopping_list.pop(remove_index)
+                        print("Item successfully removed.")
+                    else:
+                        print("Invalid number. Please enter a valid item number.")
+                
             except ValueError:
                 print("Please enter a valid number.")
 
         elif choice == '3':
             # Display the shopping list
-            if not len(shopping_list) == 0:
+            if shopping_list:
                 display_items()
             else:
                 print("Item list is empty.")
-                break
+                continue
             
         elif choice == '4':
             print("Goodbye!")
@@ -58,5 +62,5 @@ def main():
             print("Invalid choice. Please try again.")
 
 if __name__ == "__main__":
-    shopping_list = []
+    
     main()
